1/25 모델1방식 삭제 완료 admin만 수정 삭제 가능하게 함

+++ src/main/webapp/Notice/noticeView.jsp
@@ -53,10 +53,16 @@ function deletePost() {
         </tr>
         <tr>
             <td colspan="4" align="center">
-
+	
+				<%
+				if(session.getAttribute("UserId").equals("admin")){
+				%>
 				<button type="button" onclick="location.href='../Notice/Edit.do?idx=${dto.idx}';">
 					수정하기</button>
 				<button type="button" onclick="location.href='../Notice/Delete.do?flag=${dto.flag }&idx=${dto.idx}';">삭제하기</button> 
+				<%
+				}
+				%>
                 <button type="button" onclick="location.href='../Notice/List.do?flag=${dto.flag}';">
                     목록 보기
                 </button>
